requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
selenium==4.31.0
soupsieve==2.6
urllib3==2.4.0
webdriver-manager==4.0.2
python-dotenv==1.1.0
packaging==24.2
django>=4.2
slack_sdk>=3.27.0
gunicorn>=20.1.0
psycopg2-binary>=2.9
